[后端] feat(app): 添加应用相关功能- 新增应用添加、更新、删除等接口 - 实现应用列表分页查询功能 - 添加应用详情获取接口 - 实现管理员应用管理相关功能 - 新增应用相关的数据传输对象和常量类

+++ src/main/java/in/yumi/coldwindcodeai/controller/UserController.java
@@ -170,11 +170,11 @@ public class UserController {
     public BaseResponse<Page<UserVO>> listUserVOByPage(@RequestBody UserQueryRequest userQueryRequest) {
         ThrowUtils.throwIf(userQueryRequest == null, ErrorCode.PARAMS_ERROR);
         long pageNum = userQueryRequest.getPageNum();
-        long pageSize = userQueryRequest.getPageSize();
+        int pageSize = (int) userQueryRequest.getPageSize();
         Page<User> userPage = userService.page(Page.of(pageNum, pageSize),
                 userService.getQueryWrapper(userQueryRequest));
         // 数据脱敏
-        Page<UserVO> userVOPage = new Page<>(pageNum, pageSize, userPage.getTotalRow());
+        Page<UserVO> userVOPage = new Page<>((int) pageNum, pageSize, userPage.getTotalRow());
         List<UserVO> userVOList = userService.getUserVOList(userPage.getRecords());
         userVOPage.setRecords(userVOList);
         return ResultUtils.success(userVOPage);
